Removed tests from GBMGridTest and created separated JUnit test dedicated to Grid.
